+++ app/SLCANUDP.hs
@@ -1,8 +1,8 @@
 module Main where
 
-import Control.Monad.IO.Class
-import Data.Default.Class (Default(def))
-import Network.CAN (MonadCAN)
+import Control.Monad.Class.MonadSay (MonadSay(say))
+import Data.Default (Default(def))
+import Network.CAN (CAN)
 import Network.SLCAN (Transport(..))
 import Network.Socket (AddrInfo(..), SocketType(Datagram))
 
@@ -36,7 +36,7 @@ main = do
         sock
         (addrAddress ourAddrinfo)
 
-      Network.SLCAN.runSLCAN
+      Network.SLCAN.withSLCAN
         (Transport_UDP sock (addrAddress targetAddrinfo))
         def
         act
@@ -44,16 +44,17 @@ main = do
     (_, _) -> error "getAddrInfo fail"
 
 act
-  :: ( MonadCAN m
-     , MonadIO m
-     )
-  => m ()
-act = do
+  :: MonadSay m
+  => CAN m
+  -> m ()
+act can = do
   Network.CAN.send
+    can
     $ Network.CAN.standardMessage
         0x7E5
         [0x4C]
 
   Control.Monad.forever
     $ Network.CAN.recv
-      >>= Control.Monad.IO.Class.liftIO . print
+        can
+      >>= say . Network.CAN.prettyCANMessage