+++ src/Network/CAN.hs
@@ -1,7 +1,24 @@
 module Network.CAN
-  ( module Network.CAN.Class
+  ( CAN(..)
+  , send
+  , recv
   , module Network.CAN.Types
   ) where
 
-import Network.CAN.Class
 import Network.CAN.Types
+
+data CAN m = CAN
+  { canSend :: CANMessage -> m ()
+  , canRecv :: m CANMessage
+  }
+
+send
+  :: CAN m
+  -> CANMessage
+  -> m ()
+send = canSend
+
+recv
+  :: CAN m
+  -> m CANMessage
+recv = canRecv